% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_intent_mech_regex}
\alias{icd10cm_intent_mech_regex}
\title{icd10cm injury matrix by intent and mechanism combined.}
\format{Data frame}
\source{
Grouped from injury_matrix_all.
}
\usage{
icd10cm_intent_mech_regex
}
\description{
Dataset of 92 rows and 4 variables.
}
\examples{
library(dplyr)
sample_n(icd10cm_intent_mech_regex, 10)
}
\keyword{datasets}
