% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_intent_regex}
\alias{icd10cm_intent_regex}
\title{icd10cm injury matrix by intent only.}
\format{Data frame}
\source{
Grouped from injury_matrix_all
}
\usage{
icd10cm_intent_regex
}
\description{
Dataset of 5 rows and 3 variables.
}
\examples{
icd10cm_intent_regex
}
\keyword{datasets}
