% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{injury_matrix_all}
\alias{injury_matrix_all}
\title{The complete icd10cm injury matrix.}
\format{Data frame}
\source{
\url{https://www.cdc.gov/injury/wisqars/dataandstats.html}
}
\usage{
injury_matrix_all
}
\description{
Dataset of 3,602 rows and 5 variables.
formatted from the original
}
\examples{
library(dplyr)
sample_n(injury_matrix_all, 10)
}
\keyword{datasets}
