% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_intent_mechanism.R
\name{matrix_intent_mechanism}
\alias{matrix_intent_mechanism}
\title{Add Selected Intents mechanisms combined}
\usage{
matrix_intent_mechanism(data, inj_col, ...)
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}

\item{...}{keyword list}
}
\value{
return the input with additional combined intent and mechanism variables
}
\description{
Add Selected Intents mechanisms combined
}
\examples{
library(dplyr)
library(purrr)
dat <- data.frame(
  d1 = c("T63023", "X92821", "X99100", "T360x"),
  d2 = c("T65823", "Y030x0", "T17200", "V0100x")
)

}
\seealso{
\code{\link{matched_intent_mechanism}} for a more efficient approach

dat %>% matrix_intent_mechanism(inj_col = c(1, 2), "untintentional", "assault")
}
